requirements.txt
# Core ML and Data Processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2

# Google Gemini API
google-generativeai==0.3.2

# Web Framework
flask==2.3.3
flask-cors==4.0.0

# Text Processing
nltk==3.8.1
spacy==3.6.1
textblob==0.17.1

# Document Processing
PyPDF2==3.0.1
python-docx==0.8.11

# Environment Management
python-dotenv==1.0.0

# Logging and Utilities
loguru==0.7.2
requests==2.31.0

# Advanced NLP for enhanced semantic matching
transformers==4.33.2
torch==2.0.1
sentence-transformers==2.2.2

# Development Dependencies
pytest==7.4.2
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0